💾Updated infraction-logic to check for usertype before accessing

+++ Smash_Combos.Core/Cqrs/Infractions/DeleteInfraction/DeleteInfractionRequestHandler.cs
@@ -2,6 +2,7 @@
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Smash_Combos.Core.Services;
+using Smash_Combos.Domain.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -24,20 +25,37 @@ namespace Smash_Combos.Core.Cqrs.Infractions.DeleteInfraction
 
         public async Task<DeleteInfractionResponse> Handle(DeleteInfractionRequest request, CancellationToken cancellationToken)
         {
-            var infraction = await _dbContext.Infractions
-                                .Include(infraction => infraction.User)
-                                .Include(infraction => infraction.Moderator)
-                                .Where(infraction => infraction.Id == request.InfractionId && infraction.User.Id == request.UserId)
-                                .FirstOrDefaultAsync();
-            if (infraction == null)
+            User user = null;
+            try
             {
-                return new DeleteInfractionResponse { Success = false };
+                user = await _dbContext.Users.Where(user => user.Id == request.ModeratorId).SingleOrDefaultAsync();
+            }
+            catch (InvalidOperationException)
+            {
+                return new DeleteInfractionResponse { ResponseStatus = ResponseStatus.Error, ResponseMessage = "Multiple Users with same Id found" };
             }
 
-            _dbContext.Infractions.Remove(infraction);
-            await _dbContext.SaveChangesAsync(CancellationToken.None);
+            if(user == null)
+                return new DeleteInfractionResponse { ResponseStatus = ResponseStatus.BadRequest, ResponseMessage = "User does not exist" };
 
-            return new DeleteInfractionResponse { Success = true, Infraction = _mapper.Map<InfractionDto>(infraction) };
+            if(user.UserType == UserType.Moderator || user.UserType == UserType.Admin)
+            {
+                var infraction = await _dbContext.Infractions
+                                .Where(infraction => infraction.Id == request.InfractionId)
+                                .FirstOrDefaultAsync();
+
+                if (infraction == null)
+                    return new DeleteInfractionResponse { ResponseStatus = ResponseStatus.NotFound, ResponseMessage = "Infraction not found" };
+
+                _dbContext.Infractions.Remove(infraction);
+                await _dbContext.SaveChangesAsync(CancellationToken.None);
+
+                return new DeleteInfractionResponse { Data = _mapper.Map<InfractionDto>(infraction), ResponseStatus = ResponseStatus.Ok, ResponseMessage = "Infraction deleted" };
+            }
+            else
+            {
+                return new DeleteInfractionResponse { ResponseStatus = ResponseStatus.NotFound, ResponseMessage = "Not authorized to delete infractions" };
+            }
         }
     }
 }
